[wizard] add formatting helpers to get rid of \n
Smart formatting helpers 'paragraph' and 'section' added to
make formatting more ruby like instead of littering strings with
newlines.

Each section of text can specify a top and bottom margin which
represents how many newlines are expected before and after
the text. The formatter keeps track of the number of blank
newlines it last output so that if one section requests 1 bottom
margin and the next section requests 1 top margin, only one newline
is output.

If a section requests a greater number of newlines than an ajacent
section the larger number wins

If a negitive number is give for the bottom margin it is assumed
that another piece of code output untracked newlines so the
formatter takes this into account