only run wizard if a valid command is issued
This patch tightens up when the wizard is run.  If a help arg or
switch is sent the wizard won't run.  If an invalid command is
sent the wizard won't run.  If setup is called and it is the
first time through we don't tell you we are going to run
rhc setup at the end of the wizard.