Bug 915255 - Track auth preference in the config
During 'rhc setup', the user's choice to create an auth token is relevant to later interactions.  If a user declines tokens then later interactions should not recreate a token on disk, only prompt the user to login with their password.  This commit adds 'use_authorization_tokens' to the config and is a remembered value based on 3 states:

  1) did the user pass --no-create-token to the setup command?
  2) did a token get generated when the user hit 'yes'
  3) did the user pass a token to setup via --token?

The value is stored and remembered for future reference when a token expires