Bug 927425 - The use_authorization_tokens flag should be strict
Currently stored tokens are used if use_authorization_tokens is nil or false, or if the user passes --rhlogin and --password.  In both of these cases, the token should be ignored.

Removed token context helper and replaced it with a helper method "token_for_user" which is called in the commands as needed.  Also, RHC::Auth::Token is not used unless use_authorization_tokens is true or --token is passed, AND a user/password isn't provided.