Allow per-project configuration file.
RHC tools now look for the global (/etc) and local (~/.opemshift)
directories for the express.conf file.

In case when you run Origin on your machine, you need to use different
libra_server. This could be solved by using --config attribute, but
that is not really 'user-friendly' ;-)

This patch allows you to place the config file into:

$PROJECT/.openshift/express.conf

So you can overide the libra_server just for one project you want to test
on origin.