Using headers with RHC
If you were looking for a method which would allow you to use headers with the RHC tool , the configuration of a global option might be a clever choice. In order to do that , adding some lines of code into the file lib/rhc/helpers.rb is a must.

Find out a suggestion for the changes to make below!