Bug 1216206 - Always_auth should be able to be passed to 'setup'
https://bugzilla.redhat.com/show_bug.cgi?id=1216206
The always_auth option could be configured in the rhc configuration prior to running 'rhc setup' to enable always authenticating to openshift api endpoints. This change allows the '--always-auth' option to be passed in with 'rhc setup' to ensure every request is authenticated during setup and that 'always_auth=true' is added to the configuration file. This eliminates the need to manually modify the 'always_auth' directive in the configuration file before running setup.

This fix also allows the '--always-auth' option to be passed globally, with any rhc command.