Allow an ssh executable to be specified globally
Bug 1177753
https://bugzilla.redhat.com/show_bug.cgi?id=1177753

Previously, a `--ssh` option could be passed to a few subcommands to use a seperate ssh executable other than that found in the default PATH. After this commit, this option is now a global options used in all commands utilizing ssh. A `ssh` option is now available in configuration that allows this global option to be set persistently.

The snapshot-restore, snapshot-save, and tail commands will now use the an ssh executable if specified. If no ssh executable is specified, the Net::SSH implementations will continue to be used.

The port-forward command will not attempt to actually forward ports with a custom ssh executable. This is due to how errors reported from a custom ssh executable are unpredictable. If one or more of the port-forwards fails, the error reported is not predictable enough to script a workaround or helpful error message. If a custom ssh executable is specified when port-forward is called, rhc will use the custom ssh executable to gather the ports and services that will need to be forwarded from the application. It will then use this information to construct a command that the user can run to forward the ports themselves. The command structure assumes that the ssh executable uses the same flags as openSSH.

The scp command will not attempt to use any ssh executable. Instead, a helpful error will be reported with an scp command. The user can use the command to upload/download the file themselves. If the client is on windows, no scp command is provided since we cannot assume that the client has an scp executable available. Instead, a helpful error message will be provided instructing the user to investigate 3rd-party alternatives such as FileZilla and WinSCP.