rhc.gemspec: limits cucumber dependency to make code 1.8.7 friendly
Updates gemspec to limit cucumber to 1.3.20 or earlier, which in turn limits cucumber's depency of gherkin to gherkin 2.12 or earlier.  Gherkin3 requires ruby 1.9.3 or greater, whereas RHC requires 1.8.7 or later.  It also removes the cucumber-core dependency, which requires ruby 1.9.3.

Updates split_path function to different usage of CSV::parse_line depending on ruby version.  The split_path function now works in ruby 1.8.7 as well as 1.9.3 or newer.