Inform user of the need to scale up when enabling HA on a scaled application
Bug 1256952
https://bugzilla.redhat.com/show_bug.cgi?id=1256952

When an application with more than one gear ha HA enabled, a short message will now be given stating the possible need to scale the application up to actually receive the second haproxy instance.

In addition to a adding a new test, the `app_spec.rb` now requires 'net/ssh/multi' to allow the spec to be run on its own outside of the RHC spec test suite. This has been resolved.