Deobfuscated dns_propagated
Rework the `dns_propagated` method to remove the unused `found` variable
and to use a more explicit `if !host_found` instead of a less obvious
`or` (which could be easily misunderstood to be operating on expressions
rather than on statements).

This commit will also avoid the 5-second sleep if the DNS entry already
exists in the hosts file.