Bug 1163050 - Fix exception using rhc app ssh --ssh with custom path
The rhc app ssh --ssh with custom and exist path fails due to lacking
of checking whether the path is a readable file or not.

This commit modifies check_ssh_executable? method to check if the path
is a readable file or not and it no longer accepted directory as a
valid input. A exception is raised in case of unreadable/non existent
file or directory is provided.

Bug 1163050
Link <https://bugzilla.redhat.com/show_bug.cgi?id=1163050>

Signed-off-by: Vu Dinh <vdinh@redhat.com>