Include ssh url in example port-forward command
When port forwarding with a custom ssh executable and the port collection fails, an example command will be provided to the user so they can manually forward the ports. This command was previously incomplete, as it did not include the ssh url of the application. This commit completes the example command.